+++ Strata/Source/Building_StairsDown.cs
@@ -114,6 +114,8 @@ namespace Strata
         {
             base.OnEntered(pawn);
             StrataPortalUtility.TransferHaulDesignation(this, pawn);
+            StrataPortalUtility.TryDeliverConstructionCargo(pawn);
+            DraftedPortalPathing.NotifyPortalArrival(pawn);
             MapComponent_RaidPursuit.NotifyPortalArrival(pawn, pawn.MapHeld);
         }