+++ Strata/Source/Building_StairsUp.cs
@@ -123,6 +123,8 @@ namespace Strata
         {
             base.OnEntered(pawn);
             StrataPortalUtility.TransferHaulDesignation(this, pawn);
+            StrataPortalUtility.TryDeliverConstructionCargo(pawn);
+            DraftedPortalPathing.NotifyPortalArrival(pawn);
             MapComponent_RaidPursuit.NotifyPortalArrival(pawn, pawn.MapHeld);
         }